+++ src/commonMain/kotlin/com/ashampoo/xmp/XMPAliasInfo.kt
@@ -6,7 +6,7 @@
 // NOTICE:  Adobe permits you to use, modify, and distribute this file in accordance with the terms
 // of the Adobe license agreement accompanying it.
 // =================================================================================================
-package com.ashampoo.xmp.properties
+package com.ashampoo.xmp
 
 import com.ashampoo.xmp.options.AliasOptions
