+++ src/commonMain/kotlin/com/ashampoo/xmp/XMPIterator.kt
@@ -15,9 +15,8 @@ import com.ashampoo.xmp.internal.XMPNodeUtils.findNode
 import com.ashampoo.xmp.internal.XMPNodeUtils.findSchemaNode
 import com.ashampoo.xmp.options.IteratorOptions
 import com.ashampoo.xmp.options.PropertyOptions
-import com.ashampoo.xmp.properties.XMPPropertyInfo
-import com.ashampoo.xmp.xpath.XMPPath
-import com.ashampoo.xmp.xpath.XMPPathParser.expandXPath
+import com.ashampoo.xmp.internal.XMPPath
+import com.ashampoo.xmp.internal.XMPPathParser.expandXPath
 
 /**
  * Interface for the `XMPMeta` iteration services.
