+++ src/commonMain/kotlin/com/ashampoo/xmp/internal/XMPNodeUtils.kt
@@ -17,8 +17,6 @@ import com.ashampoo.xmp.internal.Utils.splitNameAndValue
 import com.ashampoo.xmp.internal.XMPUtils.encodeBase64
 import com.ashampoo.xmp.options.AliasOptions
 import com.ashampoo.xmp.options.PropertyOptions
-import com.ashampoo.xmp.xpath.XMPPath
-import com.ashampoo.xmp.xpath.XMPPathSegment
 
 /**
  * Utilities for `XMPNode`.
