+++ src/commonMain/kotlin/com/ashampoo/xmp/internal/XMPPathParser.kt
@@ -6,12 +6,10 @@
 // NOTICE:  Adobe permits you to use, modify, and distribute this file in accordance with the terms
 // of the Adobe license agreement accompanying it.
 // =================================================================================================
-package com.ashampoo.xmp.xpath
+package com.ashampoo.xmp.internal
 
 import com.ashampoo.xmp.XMPException
 import com.ashampoo.xmp.XMPMetaFactory.schemaRegistry
-import com.ashampoo.xmp.internal.Utils
-import com.ashampoo.xmp.internal.XMPErrorConst
 
 /**
  * Parser for XMP XPaths.
